+++ helper/helper_test.go
@@ -1,7 +1,6 @@
 package helper
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -107,6 +106,5 @@ func TestParseMultiValuesHeader(t *testing.T) {
 		{Value: "*", Priority: 0.3},
 	}
 	result = ParseMultiValuesHeader("fr , fr-FR;q=0.8, en-US ;q=0.5, *;q=0.3, en-*;q=0.3, en;q=0.3")
-	fmt.Println(result)
 	assert.True(t, SliceEqual(expected, result))
 }